backend/requirements.txt
django==4.1.2
coverage==6.5
pylint==2.15.5
pylint-django==2.5.3
djangorestframework==3.14.0
Pillow==9.2.0
haversine==2.7.0
